# AXMARIL/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "sean",
        "version": "1.0.0",
        "title": "Seans-Python3-Flask-Rest-Boilerplate",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/api/v1"
        },
        {
            "url": "/api/v2"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/secret": {
            "post": {
                "tags": [
                    "Secret"
                ],
                "summary": "Créer un secret",
                "description": "Crée un nouveau secret",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Données du secret à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "safe_id",
                                    "secret_type",
                                    "name"
                                ],
                                "properties": {
                                    "safe_id": {
                                        "type": "string",
                                        "description": "ID du coffre-fort"
                                    },
                                    "secret_type": {
                                        "type": "string",
                                        "enum": [
                                            "credentials",
                                            "ssh",
                                            "other",
                                            "file"
                                        ]
                                    },
                                    "app_type": {
                                        "type": "string",
                                        "description": "Type d'application"
                                    },
                                    "name": {
                                        "type": "string",
                                        "description": "Nom du secret"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "Description du secret"
                                    },
                                    "secret": {
                                        "type": "object",
                                        "description": "Contenu du secret (varie selon le type)"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Secret créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "secret_id": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Token manquant ou invalide"
                    }
                }
            }
        },
        "/secret/ssh": {
            "post": {
                "tags": [
                    "Secret"
                ],
                "summary": "Créer un secret SSH",
                "description": "Crée un secret SSH avec clé privée",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "secret",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ssh_key",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Secret SSH créé avec succès"
                    }
                }
            }
        },
        "/secret/kmip": {
            "post": {
                "tags": [
                    "Secret"
                ],
                "summary": "Créer un secret KMIP",
                "description": "Crée un secret KMIP avec certificats",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "secret",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_cert",
                        "in": "formData",
                        "type": "file"
                    },
                    {
                        "name": "user_key",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Secret KMIP créé avec succès"
                    }
                }
            }
        },
        "/secret/kmip/secret_id": {
            "post": {
                "tags": [
                    "Secret"
                ],
                "summary": "Créer un secret KMIP avec ID",
                "description": "Crée un secret KMIP en utilisant un secret_id existant",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "secret_id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Secret KMIP créé avec succès"
                    }
                }
            }
        },
        "/secret/update": {
            "put": {
                "tags": [
                    "Secret"
                ],
                "summary": "Mettre à jour un secret",
                "description": "Met à jour un secret existant",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "secret_id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "secret": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secret mis à jour avec succès"
                    }
                }
            }
        },
        "/secret/ssh/changepassword": {
            "put": {
                "tags": [
                    "Secret"
                ],
                "summary": "Changer le mot de passe SSH",
                "description": "Met à jour le mot de passe d'un secret SSH",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "secret_id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mot de passe SSH mis à jour"
                    }
                }
            }
        },
        "/secret/reveal": {
            "get": {
                "tags": [
                    "Secret"
                ],
                "summary": "Révéler un secret",
                "description": "Affiche le contenu d'un secret déchiffré",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "secret_id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "reveal_key",
                        "in": "query",
                        "type": "string",
                        "default": "false"
                    },
                    {
                        "name": "can_upload",
                        "in": "query",
                        "type": "string",
                        "default": "false"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secret révélé avec succès"
                    }
                }
            }
        },
...
        "/secret/delete": {
            "delete": {
                "tags": [
                    "Secret"
                ],
                "summary": "Supprimer un secret",
                "description": "Supprime définitivement un secret",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "secret_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secret supprimé avec succès"
                    }
                }
            }
        },
        "/secret/generate_cert": {
            "post": {
                "tags": [
                    "Secret"
                ],
                "summary": "Générer un certificat",
                "description": "Génère ou modifie un certificat",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "operation_type",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "create",
                            "modify"
                        ]
                    },
                    {
                        "name": "csr_file",
                        "in": "formData",
                        "type": "file"
                    },
                    {
                        "name": "cert_secret_id",
                        "in": "formData",
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Certificat généré/modifié avec succès"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter 'Bearer <token>'"
            }
        },
...
    }
}